move /webassembly to /capabilities/webassembly
